修改部分bug

+++ game-core/game-core-service-impl/src/main/java/com/fulu/game/core/service/impl/WxTemplateMsgServiceImpl.java
@@ -193,7 +193,9 @@ public class WxTemplateMsgServiceImpl implements WxTemplateMsgService {
             pushMsgQueue.addTemplateMessage(new WxMaTemplateMessageVO(pushId, wxMaTemplateMessage));
             formIds.add(wechatFormidVO.getFormId());
         }
-        wechatFormidService.deleteFormIds(formIds.toArray(new String[]{}));
+        if(formIds.size()>0){
+            wechatFormidService.deleteFormIds(formIds.toArray(new String[]{}));
+        }
     }
